ML/Llama/requirements.txt
Flask
langchain_groq
python-dotenv
groq
Flask-Cors
gunicorn